only do ssh auth sock fix if we have one

+++ src/bash/interactive/tmux.sh
@@ -10,7 +10,9 @@ if command_exists tmuxinator; then
 fi
 
 # fix ssh agent inside tmux
-AUTH_SOCK_LINK="${HOME}/.ssh/auth_sock"
-[[ -h $AUTH_SOCK_LINK && -z $TMUX && ! "$SSH_CLIENT" =~ "127.0.0.1" ]] && unlink $AUTH_SOCK_LINK
-[[ -z $TMUX && ! "$SSH_CLIENT" =~ "127.0.0.1" ]] && ln -s $SSH_AUTH_SOCK $AUTH_SOCK_LINK
-export SSH_AUTH_SOCK=$AUTH_SOCK_LINK
+if [ -n "$SSH_AUTH_SOCK" ]; then
+    AUTH_SOCK_LINK="${HOME}/.ssh/auth_sock"
+    [[ -h $AUTH_SOCK_LINK && -z $TMUX && ! "$SSH_CLIENT" =~ "127.0.0.1" ]] && unlink $AUTH_SOCK_LINK
+    [[ -z $TMUX && ! "$SSH_CLIENT" =~ "127.0.0.1" ]] && ln -s $SSH_AUTH_SOCK $AUTH_SOCK_LINK
+    export SSH_AUTH_SOCK=$AUTH_SOCK_LINK
+fi
